fix(bti): correct Rows.db row-index root base + leading OSS50 NEXT_COMPONENT (#3002)
Both halves of the compensating pair, in one change (fixing either alone
regresses BTI clustering reads):

A. `resolve_rows_db_entry` computed the SIGNED root-delta base as
   `RowsOffset + key_length`, 2 bytes low. Cassandra 5.0.8 captures
   `basePosition` AFTER `writeWithShortLength`
   (`BtiTableWriter.IndexWriter.append`) and reads it back as
   `in.getFilePointer()` after `readWithShortLength`
   (`BtiTableReader.retrieveEntryIfAcceptable`), i.e. the already-computed
   `entry_start`. The wrong root skipped the root node entirely and lost its
   empty-separator (block 0) payload. The writer side
   (`write_trie_index_entry`) carried the identical expression and is fixed
   with it so the write->read roundtrip stays consistent and Cassandra-true.

B. The OSS50 clustering-bound encoders emitted `0x40 NEXT_COMPONENT` only
   BETWEEN components. `ClusteringComparator.ByteComparableClustering`
   (5.0.8) emits it before EACH component including the first, so a single
   int ck=8 separator is `40 80 00 00 08` — exactly what the real
   `wide_table` Rows.db trie stores. No 0x38 TERMINATOR is added (row-index
   separators are separatorGt/nudge PREFIXES, not complete clusterings).

Also: rows.rs inline tests split to rows_tests.rs (campsite rule, #1135) and
a `rows_floor_block_for_test` hook for the fixture-backed floor oracle.
